A rough draft of the grilled cheese outline
I added a description of the grilled cheese along with, lists of ingredients and preparations. I also added the step by step process of making the grilled cheese. What I need to add is a link to making a tomato soup which I will add once I make a tomato_soup.html, which will also be linked to index.html.