Skeleton of fries html
So I added the same outline for fries html, along with an image and a description I still need to chage that for the alt elment. The I need to get back to working on ingredient & utensils list and go from there.Left a comment as well for a reminder.