I fixed a typo in the class= value for the unordered list of links of recipes.
It was originally tist-of-recipes and I changed it to list-of-recipes.
Another change I added was that class value names started witha capital letter so I changed that to start of with lower case.
**That was applied to all html files**

for the css style sheet all I added was three properties and values for class formatting-alignment with centeer alignment, auto margin, & font-style Serif family.